Move structs to their own files
